+++ ukf/run_UKF.py
@@ -31,7 +31,7 @@ import time
 
 from pyquaternion import Quaternion
 
-from alex_to_andrew_paste_fixed import *
+from UKF_algorithm import *
 
 from BNO055_MAGNETOMETER_BASIC import calibrate
 from BNO055_MAGNETOMETER_BASIC import get_data
